requirements.txt
playwright==1.45.0
PyYAML==6.0.1